+++ blade-graphics/src/gles/egl.rs
@@ -468,8 +468,7 @@ impl super::Context {
                     .unwrap()
             },
         };
-        //TODO: remove old surface
-        inner.swapchain = Some(Swapchain {
+        let old_swapchain = inner.swapchain.replace(Swapchain {
             surface,
             extent: config.size,
             format,
@@ -478,6 +477,13 @@ impl super::Context {
                 crate::DisplaySync::Recent | crate::DisplaySync::Tear => 0,
             },
         });
+        if let Some(s) = old_swapchain {
+            inner
+                .egl
+                .instance
+                .destroy_surface(inner.egl.display, s.surface)
+                .unwrap();
+        }
 
         let format_desc = super::describe_texture_format(format);
         inner.egl.make_current();